refactor(core): split turn resolution operations
